chore: Cleanup for first beta release
### Changed

- APIClient class moved to `aws_data_tools.client.APIClient`
- README clean-up
- Bump version to 0.1.0-beta1
- Adds a CI config for Semantic Pull Requests
