+++ .github/semantic.yml
@@ -1,0 +1,2 @@
+titleAndCommits: true
+scopes: ["cli", "organizations"]
